backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Neo4j
neo4j==5.23.0

# LLM (OpenAI-compatible client for Modal endpoint)
openai==1.51.0

# Paper sources
arxiv==2.1.3
semanticscholar==0.8.4
pdfplumber==0.11.4
pypdf==4.3.1

# Web search
duckduckgo-search==6.3.7
tavily-python==0.5.0

# Embeddings
sentence-transformers==3.1.1

# RocketRide SDK
#rocketride==0.1.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
tenacity==9.0.0
